Migrate all tests
